Deprecate v1 tables and processors (#141)
