--- integration-tests/sdk_expected_db_output_files/events_processor/coin_event_v2/events.json
@@ -1,45 +1,0 @@
-[
-  {
-    "sequence_number": 7,
-    "creation_number": 3,
-    "account_address": "0xa3896acd8c1aac027f776bcdc2fce50886c55b15b390cae76dc291c149a8dd23",
-    "transaction_version": 6643353707,
-    "transaction_block_height": 437887148,
-    "type_": "0x1::coin::WithdrawEvent",
-    "data": {
-      "amount": "1000000000"
-    },
-    "indexed_type": "0x1::coin::WithdrawEvent",
-    "event_index": 0
-  },
-  {
-    "sequence_number": 17,
-    "creation_number": 2,
-    "account_address": "0x0bb3e0c3db6ab22cf18b8b0c6a2efd5b2304902b2a51f6338951e85938081b78",
-    "transaction_version": 6643353707,
-    "transaction_block_height": 437887148,
-    "type_": "0x1::coin::DepositEvent",
-    "data": {
-      "amount": "1000000000"
-    },
-    "indexed_type": "0x1::coin::DepositEvent",
-    "event_index": 1
-  },
-  {
-    "sequence_number": 0,
-    "creation_number": 0,
-    "account_address": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "transaction_version": 6643353707,
-    "transaction_block_height": 437887148,
-    "type_": "0x1::transaction_fee::FeeStatement",
-    "data": {
-      "io_gas_units": "6",
-      "storage_fee_octas": "0",
-      "execution_gas_units": "5",
-      "total_charge_gas_units": "11",
-      "storage_fee_refund_octas": "0"
-    },
-    "indexed_type": "0x1::transaction_fee::FeeStatement",
-    "event_index": 2
-  }
-]
