--- integration-tests/sdk_expected_db_output_files/events_processor/empty_txn_test/events.json
@@ -1,1 +1,0 @@
-[]
