--- integration-tests/sdk_expected_db_output_files/events_processor/validator_txn_missing_events/events.json
@@ -1,1 +1,0 @@
-[]
