--- integration-tests/sdk_expected_db_output_files/fungible_asset_processor/fa_activities_txn_test/coin_supply.json
@@ -1,10 +1,0 @@
-[
-  {
-    "transaction_version": 5992795934,
-    "coin_type_hash": "91ceb1308a98389691e05158b07ed5f079ab78461a6bb8d5a4054b1bb5cb8bb6",
-    "coin_type": "0x1::aptos_coin::AptosCoin",
-    "supply": "29165678689185321503",
-    "transaction_timestamp": "2024-09-16T06:58:10",
-    "transaction_epoch": 17841
-  }
-]
