--- integration-tests/sdk_expected_db_output_files/stake_processor/83883373/events.json
@@ -1,1 +1,0 @@
-[]
