--- integration-tests/sdk_expected_db_output_files/stake_processor/mainnet_proposal_vote_test/events.json
@@ -1,1 +1,0 @@
-[]
