--- integration-tests/sdk_expected_db_output_files/stake_processor/mainnet_stake_delegation_pool/events.json
@@ -1,1 +1,0 @@
-[]
