--- integration-tests/sdk_expected_db_output_files/stake_processor/stake_delegation_pool_test/events.json
@@ -1,1 +1,0 @@
-[]
