--- integration-tests/sdk_expected_db_output_files/stake_processor/stake_distribute_test/events.json
@@ -1,1 +1,0 @@
-[]
