--- integration-tests/sdk_expected_db_output_files/stake_processor/stake_gov_record_test/events.json
@@ -1,1 +1,0 @@
-[]
