+++ integration-tests/src/diff_test_helper/mod.rs
@@ -2,7 +2,6 @@ pub mod account_restoration_processor;
 pub mod account_transaction_processor;
 pub mod ans_processor;
 pub mod default_processor;
-pub mod event_processor;
 pub mod fungible_asset_processor;
 pub mod objects_processor;
 pub mod stake_processor;
