+++ integration-tests/src/models/mod.rs
@@ -2,7 +2,6 @@ pub mod account_restoration_models;
 pub mod account_transaction_models;
 pub mod ans_models;
 pub mod default_models;
-pub mod events_models;
 pub mod fa_v2_models;
 pub mod objects_models;
 pub mod stake_models;
