+++ integration-tests/src/sdk_tests/mod.rs
@@ -7,8 +7,6 @@ pub mod ans_processor_tests;
 #[cfg(test)]
 pub mod default_processor_tests;
 #[cfg(test)]
-pub mod events_processor_tests;
-#[cfg(test)]
 pub mod fungible_asset_processor_tests;
 #[cfg(test)]
 pub mod objects_processor_tests;
