+++ processor/src/config/indexer_processor_config.rs
@@ -21,7 +21,6 @@ use crate::{
         account_restoration::account_restoration_processor::AccountRestorationProcessor,
         account_transactions::account_transactions_processor::AccountTransactionsProcessor,
         ans::ans_processor::AnsProcessor, default::default_processor::DefaultProcessor,
-        events::events_processor::EventsProcessor,
         fungible_asset::fungible_asset_processor::FungibleAssetProcessor,
         gas_fees::gas_fee_processor::GasFeeProcessor,
         monitoring::monitoring_processor::MonitoringProcessor,
@@ -68,10 +67,6 @@ impl RunnableConfig for IndexerProcessorConfig {
             ProcessorConfig::DefaultProcessor(_) => {
                 let default_processor = DefaultProcessor::new(self.clone()).await?;
                 default_processor.run_processor().await
-            },
-            ProcessorConfig::EventsProcessor(_) => {
-                let events_processor = EventsProcessor::new(self.clone()).await?;
-                events_processor.run_processor().await
             },
             ProcessorConfig::FungibleAssetProcessor(_) => {
                 let fungible_asset_processor = FungibleAssetProcessor::new(self.clone()).await?;
