+++ processor/src/parquet_processors/parquet_events/mod.rs
@@ -1,2 +1,3 @@
 pub mod parquet_events_extractor;
+pub mod parquet_events_model;
 pub mod parquet_events_processor;
