+++ processor/src/parquet_processors/parquet_events/parquet_events_extractor.rs
@@ -1,9 +1,9 @@
 use crate::{
     parquet_processors::{
-        parquet_utils::util::add_to_map_if_opted_in_for_backfill, ParquetTypeEnum,
-        ParquetTypeStructs,
+        parquet_events::parquet_events_model::{parse_events, ParquetEvent},
+        parquet_utils::util::add_to_map_if_opted_in_for_backfill,
+        ParquetTypeEnum, ParquetTypeStructs,
     },
-    processors::events::{events_model::ParquetEvent, parse_events},
     utils::table_flags::TableFlags,
 };
 use aptos_indexer_processor_sdk::{
@@ -41,7 +41,6 @@ impl Processable for ParquetEventsExtractor {
             .par_iter()
             .map(|txn| parse_events(txn, self.name().as_str()))
             .flatten()
-            .map(|e| e.into())
             .collect();
 
         let mut map: HashMap<ParquetTypeEnum, ParquetTypeStructs> = HashMap::new();
