+++ processor/src/processors/fungible_asset/coin_models/mod.rs
@@ -1,8 +1,4 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod coin_activities;
-pub mod coin_balances;
-pub mod coin_infos;
-pub mod coin_supply;
 pub mod coin_utils;
