+++ processor/src/processors/fungible_asset/fungible_asset_extractor.rs
@@ -1,5 +1,4 @@
 use crate::processors::fungible_asset::{
-    coin_models::coin_supply::CoinSupply,
     fungible_asset_models::{
         v2_fungible_asset_activities::PostgresFungibleAssetActivity,
         v2_fungible_asset_balances::{
@@ -71,7 +70,6 @@ impl Processable for FungibleAssetExtractor {
             Vec<PostgresCurrentUnifiedFungibleAssetBalance>,
             Vec<PostgresCurrentUnifiedFungibleAssetBalance>,
         ),
-        Vec<CoinSupply>,
         Vec<PostgresFungibleAssetToCoinMapping>,
     );
     type RunType = AsyncRunType;
@@ -89,7 +87,6 @@ impl Processable for FungibleAssetExtractor {
                     Vec<PostgresCurrentUnifiedFungibleAssetBalance>,
                     Vec<PostgresCurrentUnifiedFungibleAssetBalance>,
                 ),
-                Vec<CoinSupply>,
                 Vec<PostgresFungibleAssetToCoinMapping>,
             )>,
         >,
@@ -103,7 +100,6 @@ impl Processable for FungibleAssetExtractor {
             raw_fungible_asset_metadata,
             raw_fungible_asset_balances,
             (raw_current_unified_fab_v1, raw_current_unified_fab_v2),
-            coin_supply,
             fa_to_coin_mappings,
         ) = parse_v2_coin(&transactions.data, Some(&self.fa_to_coin_mapping)).await;
 
@@ -150,7 +146,6 @@ impl Processable for FungibleAssetExtractor {
                     postgres_current_unified_fab_v1,
                     postgres_current_unified_fab_v2,
                 ),
-                coin_supply,
                 postgres_fa_to_coin_mappings,
             ),
             metadata: transactions.metadata,
